bb-helper: Add MockSd
- Used for testing SD Card flashing.
