+++ bb-flasher-sd/src/bootfs_update.rs
@@ -66,16 +66,10 @@ where
 
 #[cfg(test)]
 mod test {
-    use std::io::{self, SeekFrom};
-
-    use fscommon::StreamSlice;
-    use mbrman::{CHS, MBR, MBRPartitionEntry};
+    use std::io;
+    use bb_helper::mock_sd::MockSd;
 
     use super::*;
-
-    const DISK_SIZE: u64 = 128 * 1024 * 1024; // 128 MiB
-    const SECTOR_SIZE: u32 = 512;
-    const FIRST_LBA: u32 = 2048;
 
     #[derive(Debug, Clone)]
     struct MockArchive(Vec<(Box<str>, Option<Vec<u8>>)>);
@@ -131,107 +125,6 @@ mod test {
         }
     }
 
-    #[derive(Debug)]
-    struct MockSd {
-        file: tempfile::NamedTempFile,
-        fail: CancellationToken,
-    }
-
-    impl MockSd {
-        fn new() -> Self {
-            let mut img = tempfile::NamedTempFile::new().unwrap();
-
-            img.as_file().set_len(DISK_SIZE).unwrap();
-
-            let mut mbr = MBR::new_from(&mut img, SECTOR_SIZE, [0x12, 0x34, 0x56, 0x78]).unwrap();
-
-            let total_sectors = (DISK_SIZE / SECTOR_SIZE as u64) as u32;
-            let num_sectors = total_sectors - FIRST_LBA;
-
-            mbr[1] = MBRPartitionEntry {
-                boot: 0x80,
-                first_chs: CHS::empty(),
-                sys: 0x0C, // FAT32 (LBA)
-                last_chs: CHS::empty(),
-                starting_lba: FIRST_LBA,
-                sectors: num_sectors,
-            };
-
-            mbr.write_into(&mut img).unwrap();
-
-            let partition_offset = FIRST_LBA as u64 * SECTOR_SIZE as u64;
-            let partition_size = num_sectors as u64 * SECTOR_SIZE as u64;
-
-            {
-                let mut partition = img.reopen().unwrap();
-
-                partition.seek(SeekFrom::Start(partition_offset)).unwrap();
-
-                let mut partition =
-                    StreamSlice::new(partition, partition_offset, partition_size).unwrap();
-
-                fatfs::format_volume(
-                    &mut partition,
-                    fatfs::FormatVolumeOptions::new()
-                        .fat_type(fatfs::FatType::Fat32)
-                        .volume_label(*b"BOOT       "),
-                )
-                .unwrap();
-            }
-
-            img.rewind().unwrap();
-
-            Self {
-                file: img,
-                fail: CancellationToken::default(),
-            }
-        }
-    }
-
-    impl Write for MockSd {
-        fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
-            if self.fail.is_cancelled() {
-                Err(io::Error::new(io::ErrorKind::QuotaExceeded, "Fail"))
-            } else {
-                self.file.write(buf)
-            }
-        }
-
-        fn flush(&mut self) -> io::Result<()> {
-            if self.fail.is_cancelled() {
-                Err(io::Error::new(io::ErrorKind::QuotaExceeded, "Fail"))
-            } else {
-                self.file.flush()
-            }
-        }
-    }
-
-    impl Seek for MockSd {
-        fn seek(&mut self, pos: SeekFrom) -> io::Result<u64> {
-            if self.fail.is_cancelled() {
-                Err(io::Error::new(io::ErrorKind::QuotaExceeded, "Fail"))
-            } else {
-                self.file.seek(pos)
-            }
-        }
-    }
-
-    impl Read for MockSd {
-        fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-            if self.fail.is_cancelled() {
-                Err(io::Error::new(io::ErrorKind::QuotaExceeded, "Fail"))
-            } else {
-                self.file.read(buf)
-            }
-        }
-    }
-
-    impl Eject for MockSd {
-        fn eject(self) -> io::Result<()> {
-            self.file.as_file().sync_all()
-        }
-    }
-
     #[test]
     fn basic() {
         let iter = MockArchive::default();
@@ -271,7 +164,7 @@ mod test {
         let iter_clone = iter.clone();
         flash(
             move || Ok(iter_clone),
-            crate::Destination::File(sd.file.path().into()),
+            crate::Destination::File(sd.path().into()),
             None,
         )
         .unwrap();
@@ -335,7 +228,7 @@ mod test {
         let mut sd = MockSd::new();
 
         // Break the device immediately before passing it in
-        drop(sd.fail.drop_guard());
+        drop(sd.fail_token().drop_guard());
 
         let result = internal((&mut iter).into_iter(), &mut sd, None);
 
@@ -348,7 +241,7 @@ mod test {
     #[test]
     fn test_mid_flight_storage_failure() {
         let mut sd = MockSd::new();
-        let fail_handle = sd.fail.drop_guard();
+        let fail_handle = sd.fail_token().drop_guard();
         let mut archive = MockArchive::default();
 
         // Two-way synchronization channels just to orchestrate the steps
